+++ atsfooters/src/atsfooters_internal.cpp
@@ -113,6 +113,9 @@ size_t record_footer_block_size(ats_board_type board_type,
         case ats_board_type::ats9350:
         case ats_board_type::ats9351:
             return 64;
+        case ats_board_type::ats9352:
+        case ats_board_type::ats9353:
+            return 32;
         default:
             return 128;
         }
@@ -256,12 +259,14 @@ get_internal_footer_locations(ats_footer_configuration configuration,
             for (size_t c = 0; c < active_channel_count; c++) {
                 for (size_t s = 0;
                      s < samples_per_footer / active_channel_count; s++) {
-                    const size_t source_offset = (size_t)(
-                        buf * buffer_stride_bytes + r * record_stride_bytes
-                        + c * channel_stride_bytes
-                        + (samples_per_record
-                           - footer_block_size_samples_per_channel + s)
-                              * sample_stride_bytes);
+                    const size_t source_offset
+                        = (size_t)(buf * buffer_stride_bytes
+                                   + r * record_stride_bytes
+                                   + c * channel_stride_bytes
+                                   + (samples_per_record
+                                      - footer_block_size_samples_per_channel
+                                      + s)
+                                         * sample_stride_bytes);
                     output[footer].parts.push_back({
                         source_offset,
                         bytes_per_sample,
@@ -278,10 +283,12 @@ get_internal_footer_locations(ats_footer_configuration configuration,
             break;
         case record_footer_embedding::channel_data_one_per_channel:
             for (size_t s = 0; s < samples_per_footer; s++) {
-                const size_t source_offset = (size_t)(
-                    buf * buffer_stride_bytes + r * record_stride_bytes
-                    + (samples_per_record - footer_block_size_samples + s)
-                          * sample_stride_bytes);
+                const size_t source_offset
+                    = (size_t)(buf * buffer_stride_bytes
+                               + r * record_stride_bytes
+                               + (samples_per_record - footer_block_size_samples
+                                  + s)
+                                     * sample_stride_bytes);
                 output[footer].parts.push_back(
                     {source_offset, bytes_per_sample});
             }