+++ python-example/example-atsfooters.py
@@ -10,11 +10,11 @@ from atsfooters import (
 
 
 def main():
-    data = np.fromfile("data-ats9350-fft-1ch-2112.bin", np.uint8)
-    config = FooterConfiguration(BoardType.ats9350, DataDomain.frequency, 1,
-                                 DataLayout.sample_interleaved, 2112, 2, False)
+    data = np.fromfile("data.bin", np.uint8)
+    config = FooterConfiguration(BoardType.ats9352, DataDomain.frequency, 2,
+                                 DataLayout.record_interleaved, 4128, 2, False)
     lib = ATSFooters()
-    footers = lib.parse_type_0(data, config, 4)
+    footers = lib.parse_type_1(data, config, 4)
     for f in footers:
         print(f)