+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to `ats-footers` are documented here.
 ### Added
 - Python wrapper
 
+### Fixed
+- Issue parsing ATS9352 2-channel FFT footers
+
 ## [0.1.0] - 2021-06-25
 ### Added
 - Initial version